Pick the subtitle language from settings
Currently only the first the subtitle/caption is picked. Adding code to get the language json file - based on the app settings, and if it's the 2 letter language then change it to 5 letter something like "es_ES", "it_IT" and use it to detect the subtitle/caption that would be downloaded. It's not perfect but if we can't convert from 2 to 5 letter code the json file can be named as "pt_BR" or "en_US". Another option is to no try to convert from 2 to 5 letter code but change the json filenames to 5 letter code.